fix erros compilation
