+++ content/guides/github-action-integration/code/EcommerceSampleTS.ts
